test projet casse

+++ includes/miniRT_solids.h
@@ -6,7 +6,7 @@
 /*   By: aykrifa <aykrifa@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/10 10:23:03 by aykrifa           #+#    #+#             */
-/*   Updated: 2025/07/10 13:51:00 by aykrifa          ###   ########.fr       */
+/*   Updated: 2025/07/11 11:29:40 by aykrifa          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,9 +17,9 @@
 
 typedef struct s_rgb
 {
-	float	r;
-	float	g;
-	float	b;
+	double	r;
+	double	g;
+	double	b;
 }	t_rgb;
 
 typedef struct s_base
@@ -43,11 +43,11 @@ typedef struct s_sp
 {
 	t_type	type;
 	t_vect	center;
-	float	diameter;
+	double	diameter;
 	t_rgb	color;
-	float	reflexion;
+	double	reflexion;
 
-	float	radius;
+	double	radius;
 }	t_sp;
 
 typedef struct s_pl
@@ -56,7 +56,7 @@ typedef struct s_pl
 	t_vect	point;
 	t_vect	normal_n;
 	t_rgb	color;
-	float	reflexion;
+	double	reflexion;
 
 	double	d;
 }	t_pl;
@@ -66,14 +66,14 @@ typedef struct s_cy
 	t_type	type;
 	t_vect	center;
 	t_vect	axis_n;
-	float	diameter;
-	float	height;
+	double	diameter;
+	double	height;
 	t_rgb	color;
-	float	reflexion;
+	double	reflexion;
 
 	t_base	base;
-	float	radius;
-	float	semi_height;
+	double	radius;
+	double	semi_height;
 	t_vect	top;
 	t_vect	bottom;
 	double	dt;
@@ -85,14 +85,14 @@ typedef struct s_co
 	t_type	type;
 	t_vect	center;
 	t_vect	axis_n;
-	float	diameter;
-	float	height;
+	double	diameter;
+	double	height;
 	t_rgb	color;
-	float	reflexion;
+	double	reflexion;
 
 	t_vect	apex;
-	float	radius;
-	float	k;
+	double	radius;
+	double	k;
 }	t_co;
 
 #endif
