+++ srcs/init/constructors_utils.c
@@ -6,7 +6,7 @@
 /*   By: cbordeau <cbordeau@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/09 08:52:36 by cbordeau          #+#    #+#             */
-/*   Updated: 2025/07/05 09:34:03 by aykrifa          ###   ########.fr       */
+/*   Updated: 2025/07/11 11:33:59 by aykrifa          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,21 +32,21 @@ int	fill_rgb(char *str, t_rgb *color, char *error)
 	if (!str)
 		return (print_error(ARGS, error));
 	cursor = str;
-	color->r = (float)ft_atoi(cursor);
+	color->r = (double)ft_atoi(cursor);
 	// if (color->r > 1 || color->r < 0)
 	// 	return (print_error(COLOR, error));
 	cursor = get_next_number(cursor);
 	if (*cursor != ',')
 		return (print_error(ARGS, error));
 	cursor++;
-	color->g = (float)ft_atoi(cursor);
+	color->g = (double)ft_atoi(cursor);
 	// if (color->g > 1 || color->g < 0)
 	// 	return (print_error(COLOR, error));
 	cursor = get_next_number(cursor);
 	if (*cursor != ',')
 		return (print_error(ARGS, error));
 	cursor++;
-	color->b = (float)ft_atoi(cursor);
+	color->b = (double)ft_atoi(cursor);
 	// if (color->b > 1 || color->b < 0)
 	// 	return (print_error(COLOR, error));
 	return (0);
